Modified longClickListener to return result of the embedded listener.
In appendTextActionToBar a longClickListener is set for the button. This
embeds a longClickListener object, which will perform an action
depending on the action identifier it got on construction.

In case no action is designated for that action identifier, the listener
will return `false`. This result should be returned by the enclosing
longClickListener object. This is evaluated by View::performLongClick().
In case no action was performed, the tooltip may be shown.