markdown: Add admonition extension, by @fgtham @gsantner (PR #1584)
* Add support for flexmark admonition extension https://github.com/vsch/flexmark-java/wiki/Admonition-Extension)
which allows for easy creation of colored boxes

```
!!! note "Note"
    This is the note body. Full markdown content is allowed.
```
If we use the unmodified upstream javascript, the collapse toggle
doesn't work. The provided function needs to be wrapped in an event
listener.

In order to facilitate maintenance, make only minimal changes to the
upstream JS file and document them.

When an admonition box is used in a list, the list item's text-indent
attribute is applied to the boxed content as well. Thus, the text is not
contained in the box any longer. This commit removes the text-indent
from the box.

Co-authored-by: Florian Tham <fgt@mailbox.org>
Co-authored-by: Gregor Santner <gsantner@mailbox.org>