Set HTML lang attribute
After conversion from markdown to html, the document is wrapped in a `div`
with the `lang` attribute set to the app language. For a specific
document, the language can be overwritten by specifying the `lang`
attribute in the YAML header.

This change alone will not do anything visible to the user, but now she
can easily enable hyphenation, if desired, by injecting the appropriate
CSS:

```css
html {
    word-break: break-word;
    hyphens: auto;
}
```

This might use display space more efficiently on small devices.