Automatically set HTML lang attribute, by @fgtham (PR #1623)
After conversion from markdown to html, the document is wrapped in a `div`
with the `lang` attribute set to the app language. For a specific
document, the language can be overwritten by specifying the `lang`
attribute in the YAML header. This might use display space more efficiently on small devices.This change alone will not do anything visible to the user, but now she can easily enable hyphenation, if desired, by injecting the appropriate CSS:

```css
html {
    word-break: break-word;
    hyphens: auto;
}
```

Co-authored-by: Florian Tham <fgtham@gmail.com>
Co-authored-by: Gregor Santner <gsantner@mailbox.org>