Zim: Fix link/attachment path of file selection
While editing a wiki file, Zim adds links to a target residing inside
the Notebook (where notebook.zim file is found) resolving the path as
relative to the wiki file's attachment directory (same parent/name of
the wiki file, but without the extension).  If a link's target is not
a child of the Notebook, Zim adds the link as an absolute path.

Zim always prefixes relative paths, adding './' if a link's target is
a child of a wiki file's attachment directory.

/home/user/Notebook            # Notebook directory
      |    |
      |    `- notebook.zim     # marks the root of the Notebook
      |    |
      |    `- wiki-example     # wiki file's attachment directory
      |    |  |
      |    |  `- video.mkv     # wiki-example.txt file attachment
      |    |
      |    `- wiki-example.txt # wiki file written in wiki markup
      |
      `- README                # file outside the Notebook's root

In wiki-example.txt, [[./video.mkv]] would be a link to an attachment
residing in the 'wiki-example' directory.  Linking notebook.zim would
add [[../notebook.zim]], and [[/home/user/README]] would be a link to
a file outside the Notebook.